refactor: split KitGridItem into focused modules
KitGridItem.tsx was the largest renderer god-object from the audit
(635 lines): one file holding two popover bodies, the delete and
duplicate confirmation state machines, search-highlight helpers, a
90-line voice-strip render block, and the actual grid item layout.
Extract each concern along its existing seam:

- shared/KitItemActionPopovers.tsx: the DeletePopoverContent /
  DuplicatePopoverContent bodies (moved verbatim, now exported).
- hooks/kit-management/useKitItemActions.ts: the delete/duplicate
  popover state machines, the exit animation on confirmed delete,
  and the legacy no-popover fallbacks. Exports the DuplicateKitFn /
  RequestDeleteSummaryFn types the props reference.
- KitVoiceStrip.tsx: the four-voice channel strip, including the
  search-match filename expansion and per-voice count styling.
- shared/searchHighlight.tsx: highlightMatch + stripExtension, used
  by both the alias line and the voice strip.

KitGridItem keeps the layout and the favorite-pulse state (~330
lines). Markup is unchanged; the existing KitGridItem / KitGrid /
KitGridCard suites (126 tests, covering delete/duplicate popovers,
search highlighting, and the voice strip) pass unmodified.
